# openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.0
  title: Blog API
  description: |
    This API is being used to manage the blog of **Insert Company Here**. By using this API, you can create an entry for blog, edit you blog post,
    find you blogpost, or even delete it. To start using this API, you must create an account for **Insert Company Here**.
servers:
  - url: http://localhost:8080/
    description: "Mock Server"
...
security:
  - BearerToken: []
paths:
...
  /v1/register:
    post:
      summary: "Create new account"
      description: |
        # Create new account

        Use this endpoint to create new account for this site.
        If the email already exist it will return `422`.
      operationId: doRegister
      security: []
      requestBody:
        $ref: "#/components/requestBodies/RegistrationRequest"
      tags:
        - Authentication
      responses:
        "201":
          $ref: "#/components/responses/TokenOKResponse"
        "422":
          $ref: "#/components/responses/EmailAlreadyExistErrorResponse"
...
components:
  schemas:
...
    TokenEntity:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        expiresAt:
          type: number
          description: Time since epoch
          example: 121323600
  requestBodies:
...
    RegistrationRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - email
              - name
              - password
            properties:
              name:
                type: string
                example: "Ahmad Nizar"
              email:
                type: string
                example: "ahmad@nizar.co.id"
              password:
                type: string
                example: "123456"
  responses:
...
    EmailAlreadyExistErrorResponse:
      description: Failed to register response body schema
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                type: object
                properties:
                  name:
                    type: string
                    example: FAILED_TO_REGISTER
                  message:
                    type: string
                    example: "Email already exists!"
    TokenOKResponse:
      description: Login and register body schema
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                $ref: "#/components/schemas/TokenEntity"
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Retrieve it via Authentication tagged endpoint."